refactor: Migrate MFGProblem to unified BC interface from conditions.py
Phase 1 of issue #395: Replace fdm_bc_1d dependency with unified BC.

Changes:
- Update MFGProblem imports: BoundaryConditions from conditions.py
- Use periodic_bc() factory function instead of BoundaryConditions(type="periodic")
- Update get_boundary_conditions() to return unified BC
- Update test imports to distinguish unified vs legacy BC types

The unified BC from conditions.py supports nD boundary conditions while
maintaining backward compatibility with legacy 1D BC through MFGComponents.
